% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_score.R
\name{connectivity_score}
\alias{connectivity_score}
\title{Broad CMAP gene set enrichment metrics}
\usage{
connectivity_score(data.matrix, id.up, id.down)
}
\arguments{
\item{data.matrix}{the gene expression matrix}

\item{id.up}{first gene set, usually the up-regulated genes in a cell line}

\item{id.down}{second gene set, usually the down-regulated genes in a cell
line}
}
\description{
A method for computing Broad CMAP connectivity scores. This method is a kind
of gene set enrichment analysis, it's initially designed for microarray
samples, but could also applied to RNA-seq data. To use this method, we need
a expression matrix and two corresponding gene sets.
}
\references{
\url{http://portals.broadinstitute.org/cmap/help_topics_linkified.jsp#connectivity score}
}
\seealso{
\code{\link[gCMAP]{connectivity_score}}
}

