% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_table.R
\name{form_table}
\alias{form_table}
\title{Form Contingency Table}
\usage{
form_table(set1, set2, total = NULL)
}
\arguments{
\item{set1}{the 1st set}

\item{set2}{the 2nd set}

\item{total}{the full set}
}
\description{
Construct contingency table from two sets. This step can be used to prepare
for fisher exact test.
}
\examples{
form_table(letters[1:10], letters[5:15], letters)

}

