% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_package.R
\name{has_package}
\alias{has_package}
\title{Check if a package is already installed}
\usage{
has_package(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{a character vector, each value represents a package name.}

\item{lib.loc:}{a character vector of directory names of R libraries, or
'NULL'.  The default value of 'NULL' corresponds to all
libraries currently known.'}
}
\value{
a logical vector. TRUE means the package is installed.
}
\description{
Read package list and check if a package is installed or not.
}
\examples{
has_package('stringr')

}

