% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapidx.R
\name{mapidx}
\alias{mapidx}
\title{Get matched index}
\usage{
mapidx(query, subject, na.rm = TRUE, sort = TRUE, uniq = TRUE)
}
\arguments{
\item{query, subject}{each of them is a vector}

\item{na.rm}{remove NA from result?}

\item{sort}{sort result?}

\item{uniq}{remove duplicated value?}
}
\value{
a integer vector
}
\description{
Map elements from one set to another, and return the index
}
\examples{
mapidx(c('h','e','l','l','o'), c('w','o','r','l','d'))

}
