% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageAuthor.R
\name{packageAuthor}
\alias{packageAuthor}
\title{Package Author Information}
\usage{
packageAuthor(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{a character string with the package name.}

\item{lib.loc}{a character vector of directory names of R libraries, or
'NULL'.  The default value of 'NULL' corresponds to all
libraries currently known.'}
}
\value{
a character vector, with each element represents an author name. When the
package isn't available, an empty vector would be returned.
}
\description{
Parses and returns author-associated information of a package.
}
\examples{
packageAuthor('base')

}
