% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageMatainer.R
\name{packageMaintainer}
\alias{packageMaintainer}
\title{Package Maintainer Information}
\usage{
packageMaintainer(pkg, lib.loc = NULL)
}
\arguments{
\item{pkg}{a character string with the package name.}

\item{lib.loc}{a character vector of directory names of R libraries, or
'NULL'.  The default value of 'NULL' corresponds to all
libraries currently known.'}
}
\value{
a character vector, with each element represents an maintainer name. When the
package isn't available, an empty vector would be returned.
}
\description{
Parses and returns maintainer-associated information of a package.
}
\examples{
packageMaintainer('base')

}

