% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_factor.R
\name{prime_factor}
\alias{prime_factor}
\title{Do prime factorization on integers}
\usage{
prime_factor(x, uniq = FALSE)
}
\arguments{
\item{x}{an numerical vector.}

\item{uniq}{logical; if TRUE, duplicated prime factors would be removed.}
}
\value{
a list, with length equal to the input vector.
}
\description{
Invoke 'factor' command (part of GNU coreutils)
}
\examples{
prime_factor(1001:1010)
prime_factor(65536, uniq=TRUE)

}
