% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_htseq.R
\name{read_htseq}
\alias{read_htseq}
\title{Read multiple HTSeq-count data into a matrix}
\usage{
read_htseq(files, labels = NULL, clean = TRUE)
}
\arguments{
\item{files}{HTSeq-count output files}

\item{labels}{sample labels used as colnames in count matrix}

\item{clean}{clean meta-line (start with `__`) or not}
}
\description{
HTSeq-count output a tab-delimed file, while the first column records
gene_id, and the second column records raw read count. This function merges
multiple HTSeq-count output into one matrix.
}

