% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_goana.R
\name{run_goana}
\alias{run_goana}
\title{Run gene ontology enrichment analysis}
\usage{
run_goana(de, species = "Hs", FDR = 0.01, N.min = 10)
}
\arguments{
\item{de}{a character vector of Entrez Gene IDs, or a list of such vectors, or an `MArrayLM` fit object.}

\item{species}{character string specifying the species. Possible values include "Hs" (human), "Mm" (mouse), "Rn" (rat), "Dm" (fly) or "Pt" (chimpanzee), but other values are possible if the corresponding organism package is available.}

\item{FDR}{false discovery rate cutoff for differentially expressed genes. Numeric value between 0 and 1.}

\item{N.min}{minimal number of genes in a GO term}
}
\value{
a data_frame
}
\description{
Find enriched GO terms from a list of gene entrez id or gene symbol.
}
\examples{
gene_list = c('SIGLEC1', 'ADAM33', 'PPP1R14A', 'FFAR1', 'CD22', 'DNASE2',
              'TNFSF13', 'CCND3', 'VASH1', 'RNF141', 'MED12L', 'SIGLEC6',
              'GPR160', 'POU3F1', 'AXL', 'ING1', 'EPHB4', 'SLC1A4', 'B3GNT7',
              'ATP1A2', 'SPOCK2', 'TGFBR1', 'LRRC8A', 'LTA4H', 'VASN', 'LTK',
              'CDH1')
run_goana(gene_list, FDR = 0.05)

}
\seealso{
\code{\link[limma]{goana}}
}
