% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn}
\alias{venn}
\title{Generate a plot to illustrate the relationship between two sets}
\usage{
venn(set1, set2, uniq = TRUE, xlab = "set2", ylab = "set1", ...)
}
\arguments{
\item{set1}{1st vector of values}

\item{set2}{2nd vector of values}

\item{uniq}{should duplicated values be removed?}

\item{xlab}{default title for the x axis}

\item{ylab}{default title for the y axis}

\item{...}{other parameters passed to `plot` function}
}
\description{
This function does the following things: 1) get input from two vectors; 2)
calculate the intersect between them; 3) show the relative ratio between the
intersect size and the original sets. To some extent it has the same
functionality as the venn plot.
}
\examples{
venn(sample(letters, 20), sample(letters, 25))
venn(sample(letters, 20, TRUE), sample(letters, 25, TRUE), FALSE)

}
