% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Z-score normalization}
\usage{
zscore(x, byrow = TRUE)
}
\arguments{
\item{x}{a numeric matrix}

\item{byrow}{normalize by each row of matrix?}
}
\description{
A wrapper for the function \code{\link{scale}}
}
\examples{
x <- matrix(1:16, 4, 4)
zscore(x)
zscore(x, byrow=FALSE)

}

