% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_package.R
\name{has_package}
\alias{has_package}
\title{Check if a package is already installed}
\usage{
has_package(x)
}
\arguments{
\item{x}{a character vector, each value represents a package name.}
}
\value{
a logical vector. TRUE means the package is installed.
}
\description{
Read package list and check if a package is installed or not.
}
\examples{
has_package('stringr')

}

