% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot.pca}
\alias{plot_pca}
\title{Generate a PCA plot from a gene expression matrix}
\usage{
plot_pca(count, color, shape, color.legend = "cell", shape.legend = "donor")
}
\arguments{
\item{count}{the gene expression count matrix}

\item{color}{a character vector whose length is same as samples, different
value corresponds to different point color in the plot}

\item{shape}{a character vector whose length is same as samples, different
value corresponds to different point shape in the plot}

\item{color.legend}{the legend text corresponds to point color}

\item{shape.legend}{the legend text corresponds to point shape}
}
\value{
a ggplot object
}
\description{
This function is able to log transform (use rlog function in DESeq2 package)
count data originated from a RNA-seq experiment, and draw a PCA plot to
reflect sample relationship. That's a good way to check batch effect from
gene expression matrix.
}
\examples{
count = matrix(rpois(400, 101:140), 100, 4)
colnames(count) <- paste("sample", 1:4, sep="_")
rownames(count) <- paste("gene", 1:100, sep="_")
plot_pca(count, rep(c('c1','c2'), 2), rep(c('d1', 'd2'), each=2))

}

